# Performance: ~250% faster JSON Parser